refactor: organize search, card, flyout components by features/entities
