+++ src/components/Pagination.tsx
@@ -1,4 +1,4 @@
-import { Button } from './Button';
+import { Button } from '../shared/ui/kit/Button';
 
 export function Pagination({
   hasNextPage,
