+++ src/entities/Card/ui/PinChip.tsx
@@ -1,16 +1,16 @@
 export function PinChip({
-  isSaved,
+  isSelected,
   content,
   className,
   onClick,
 }: {
-  isSaved: boolean;
+  isSelected: boolean;
   content: string;
   className?: string;
   onClick?: () => void;
 }) {
   const base = `absolute top-0 right-0 m-1 cursor-pointer items-center justify-center rounded-full bg-[var(--color-overlay)] px-2 py-1 font-medium text-[var(--color-primary)] hover:bg-[var(--color-bg)]`;
-  const visibility = isSaved
+  const visibility = isSelected
     ? 'flex bg-[var(--color-overlay)]'
     : 'hidden group-hover:flex';
   return (
