+++ src/features/search/index.ts
@@ -1,0 +1,1 @@
+export { SearchLayout } from './ui/Layout';
