+++ src/features/search/lib/getSearchTip.ts
@@ -1,0 +1,4 @@
+import { searchTips } from '../model/constants';
+
+export const getSearchTip = () =>
+  searchTips[Math.floor(Math.random() * searchTips.length)];
