+++ src/features/search/model/constants.ts
@@ -1,0 +1,13 @@
+export const LOCAL_STORAGE_KEY = '___search_history_43242';
+
+export const INITIAL_QUERY = 'Saturn';
+
+export const searchTips = [
+  'e.g. nebula, jupiter, apollo 11, mars rover...',
+  'Try: milky way, saturn rings, ISS, SpaceX ...',
+  'e.g. Falcon 9, Starship, Crew Dragon...',
+  'e.g. black hole, curiosity, galaxies...',
+  'e.g. aurora, moon landing, helix nebula...',
+  'e.g. comet, galaxy cluster, lunar eclipse...',
+  "e.g. exoplanets, pulsars, pluto's heart...",
+];
