+++ src/shared/ui/kit/Button.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../shared/lib/cn';
+import { cn } from '../../lib/cn';
 
 export const Button = ({
   content,
